% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ctxml.R
\name{create_ctxml}
\alias{create_ctxml}
\title{Creates xml document for upload to clinicaltrials.gov}
\usage{
create_ctxml()
}
\value{
A xml document
}
\description{
This function will create an xml document conforming to clinicaltrials.gov requirements for
automatic upload to the registry
}
