% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ctxml.R
\name{print_ctxml}
\alias{print_ctxml}
\title{Print xml document created using spiritR}
\usage{
print_ctxml(ctxml)
}
\arguments{
\item{ctxml}{The xml document generated by a call to create_ctxml()}
}
\description{
This function allows you to easily view the structure of the xml document
generated using the create_ctxml() function
}
