% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_ct.R
\name{upload_ct}
\alias{upload_ct}
\title{Upload an xml object to the clinicaltrials.gov registry}
\usage{
upload_ct(xmlobject, orgName, userName, passWord)
}
\arguments{
\item{xmlobject}{A xml document}

\item{orgName}{The organisation name associated with a clinicaltrials.gov
account}

\item{userName}{Username for a clinicaltrials.gov account}

\item{passWord}{Password for a clinicaltrials.gov account}
}
\value{
A message from a http post request to show that the upload was
successful or unsuccesful
}
\description{
This function will make a http POST request to upload a XML document to the
clinicaltrials.gov registry.
}
