% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_functions}
\alias{add_functions}
\alias{add_arm}
\alias{add_intervention}
\alias{add_pr_outcome}
\alias{add_sec_outcome}
\title{Add arms, interventions and outcomes to an xml document}
\usage{
add_arm(ctxml, arm_label, arm_type, arm_desc)

add_intervention(ctxml, int_type, int_name, int_desc, arm_label)

add_pr_outcome(ctxml, name, time, description)

add_sec_outcome(ctxml, name, time, description)
}
\value{
A xml nodespace
}
\description{
These functions add arms, interventions, primary and secondary outcomes to
the xml document created using the create_ctxml() function. Calls to these
functions should not be assigned to an object.
}
\details{
\itemize{
\item \code{add_arm()}: Adds an xml nodespace containing information about the arm
name, type and description.
\item \code{add_intervention()}: Adds an xml nodespace containing information about
the intervention name, type, description and arm it is associated with.
\item \code{add_pr_outcome()}: Adds an xml nodespace containing information about the
outcome name, time frame for measurement and additional descriptive details.
\item \code{add_sec_outcome()}: Adds an xml nodespace containing information about the
outcome name, time frame for measurement and additional descriptive details.
}
}
\examples{
\dontrun{
 add_arm(ctxml = ctxml, 
         arm_label = "Standard", 
         arm_type =  "Active Comparator", 
         arm_desc = "Manual upload to registry")
}
\dontrun{
add_intervention(ctxml = ctxml, 
                 int_type =  "Device", 
                 int_name = "Registry entry",
                 int_desc = "The usual way to enter to the registry",
                 arm_label = "Standard")
}
\dontrun{
add_pr_outcome(ctxml = ctxml,
               name = "correct upload",
               time = "As measured",
               description = "Insert description about the measure.")
}
\dontrun{
add_sec_outcome(ctxml = test_ctxml,
                name = "time to upload",
                time = "As measured",
                description = "Insert description about the measure.")
}
}
