% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiritR-package.R
\docType{package}
\name{spiritR-package}
\alias{spiritR}
\alias{spiritR-package}
\title{spiritR: Template for Clinical Trial Protocol}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Contains an R Markdown template for a clinical trial 
    protocol adhering to the SPIRIT statement. The SPIRIT (Standard Protocol 
    Items for Interventional Trials) statement outlines recommendations for a 
    minimum set of elements to be addressed in a  clinical trial protocol. 
    Also contains functions to create a xml document from the template and 
    upload it to clinicaltrials.gov<https://www.clinicaltrials.gov/> for 
    trial registration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/awconway/spiritR}
  \item Report bugs at \url{https://github.com/awconway/spiritR/issues}
}

}
\author{
\strong{Maintainer}: Aaron Conway \email{aaron.conway@utoronto.ca} (0000-0002-9583-8636)

}
\keyword{internal}
